# api/openapi.yaml
openapi: 3.0.3
info:
  title: Gophermart — накопительная система лояльности
  version: 1.0.0
  description: |
    HTTP API накопительной системы лояльности «Гофермарт».

    Контракт зафиксирован техническим заданием ([`ToR.md`](../ToR.md)) и является
    источником истины для обработчиков: маршруты, коды ответа и имена JSON-полей
    не меняются без явной причины и обновления этого документа.

    # Аутентификация

    Регистрация и вход возвращают токен доступа **двумя способами одновременно**:
    в заголовке ответа `Authorization` со значением `Bearer <token>` и в cookie
    `auth_token`. Техническое задание допускает любой из этих способов, а
    одновременная выдача снимает зависимость от того, какой из них читает клиент.

    Защищённые эндпоинты принимают токен из заголовка `Authorization`, а при его
    отсутствии — из cookie. Если присутствуют оба источника, проверяется значение
    из заголовка.

    Cookie помечена как `HttpOnly` и не помечена как `Secure`: сервис
    обслуживает запросы по обычному HTTP, и флаг `Secure` привёл бы к тому, что
    клиент отбросил бы cookie.

    # Сжатие

    Сервис сжимает тело ответа алгоритмом gzip, если клиент передал
    `Accept-Encoding: gzip`, и принимает сжатое тело запроса при
    `Content-Encoding: gzip`. Ответ, содержимое которого зависит от кодировки,
    содержит заголовок `Vary: Accept-Encoding`. Повреждённый gzip-поток в теле
    запроса приводит к коду `400`.

    # Денежные значения

...

    # Идентификаторы запросов
...
servers:
...
paths:
  /api/user/register:
    post:
      tags: [Пользователь]
      summary: Регистрация пользователя
      description: |
        Регистрирует пользователя по паре логин/пароль и сразу его
        аутентифицирует. Каждый логин уникален.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Пользователь зарегистрирован и аутентифицирован
          headers:
            Authorization:
              $ref: '#/components/headers/AuthorizationToken'
            Set-Cookie:
              $ref: '#/components/headers/AuthCookie'
            X-Request-Id:
              $ref: '#/components/headers/RequestID'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Логин уже занят
          content:
            text/plain:
              schema:
                type: string
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/user/login:
    post:
      tags: [Пользователь]
      summary: Аутентификация пользователя
      description: Проверяет пару логин/пароль и выпускает новый токен доступа.
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Пользователь аутентифицирован
          headers:
            Authorization:
              $ref: '#/components/headers/AuthorizationToken'
            Set-Cookie:
              $ref: '#/components/headers/AuthCookie'
            X-Request-Id:
              $ref: '#/components/headers/RequestID'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            Неверная пара логин/пароль. Ответ не позволяет отличить
            несуществующий логин от неверного пароля.
          content:
            text/plain:
              schema:
                type: string
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Токен доступа в заголовке `Authorization`. Имеет приоритет над cookie.
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Токен доступа в cookie. Используется, если заголовок `Authorization` отсутствует.
  headers:
    AuthorizationToken:
      description: |
        Выпущенный токен доступа в формате `Bearer <token>`. Возвращается вместе
        с cookie: клиент может использовать любой из способов.
      schema:
        type: string
      example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    AuthCookie:
      description: |
        Cookie `auth_token` с тем же токеном. Помечена `HttpOnly`, флаг `Secure`
        не выставляется, поскольку сервис работает по HTTP.
      schema:
        type: string
      example: auth_token=eyJhbGciOiJIUzI1NiJ9...; Path=/; HttpOnly; SameSite=Lax
    RequestID:
      description: |
        Идентификатор запроса. Совпадает со значением `X-Request-Id` из запроса,
        если клиент его передал, иначе сгенерирован сервисом.
      schema:
        type: string
      example: af01524be117dd71f44539a2117d4a44
  responses:
    BadRequest:
      description: Неверный формат запроса
      content:
        text/plain:
          schema:
            type: string
...
    PayloadTooLarge:
      description: |
        Тело запроса превышает допустимый размер. Код не предусмотрен техническим
        заданием и является защитным ограничением реализации: лимит заведомо
        превышает размер любого легитимного запроса контракта.
      content:
        text/plain:
          schema:
            type: string
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        text/plain:
          schema:
            type: string
  schemas:
    Credentials:
      type: object
      description: Пара логин/пароль.
      required: [login, password]
      properties:
        login:
          type: string
          minLength: 1
          description: Логин пользователя. Уникален, сравнивается с учётом регистра.
          example: gopher
        password:
          type: string
          minLength: 1
          maxLength: 72
          format: password
          description: |
            Пароль в открытом виде. Хранится только в виде адаптивного хеша.
            Ограничение длины продиктовано алгоритмом хеширования: bcrypt
            использует лишь первые 72 байта пароля, поэтому более длинное
            значение отклоняется, а не усекается молча.
          example: correct-horse-battery-staple
...
security:
  - bearerAuth: []
  - cookieAuth: []